perubahan pada 03 juni 2025

+++ backend/src/main/java/edu/ut/kelompokb/notaryapp/entities/Appointment.java
@@ -14,7 +14,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.PrePersist;
 import jakarta.persistence.PreUpdate;
 import jakarta.persistence.Table;
@@ -37,7 +36,7 @@ public class Appointment {
 
     private String description;
 
-    private int durationMinutes;
+    private long durationMinutes;
 
     @Enumerated(EnumType.STRING)
     private AppointmentStatus status = AppointmentStatus.PENDING;
@@ -97,11 +96,11 @@ public class Appointment {
         this.description = description;
     }
 
-    public int getDurationMinutes() {
+    public long getDurationMinutes() {
         return durationMinutes;
     }
 
-    public void setDurationMinutes(int durationMinutes) {
+    public void setDurationMinutes(long durationMinutes) {
         this.durationMinutes = durationMinutes;
     }
